Extract IssueLocation
